Delete separate plugin, added to online_monitor plugin
